// src/database/common/enum/database.enum.ts
export enum DB_TABLE {
  MEDIAL_OBJECT = 'media_object',
  ROLE = 'role',
  USER = 'user',
}

export enum ENUMBaseEntity {
  ID = 'id',
}

export enum ENUMCommonEntity {
  CREATED_AT = 'created_at',
  DELETED_AT = 'deleted_at',
  IS_DELETE = 'is_delete',
  UPDATED_AT = 'updated_at',
}

export enum ENUMTypeColumnEntity {
  TYPE_ID = 'int4',
  TYPE_DATE = 'int8',
}